EES-3679 refactoring
